Updated installation scripts and fixed weekly report bug

Closes #1

See merge request !1

+++ scripts/install.sh
@@ -1,7 +1,7 @@
 CONFFILE=/usr/share/nginx/poppet/instance/config.py
 
 apt-get update
-apt-get install nginx redis-server mysql-server mysql-client python-dev python libmysqlclient-dev python-pip git libffi-dev upstart supervisor libssl-dev libtiff5-dev libjpeg8-dev zlib1g-dev libfreetype6-dev liblcms2-dev libwebp-dev -y
+apt-get install nginx redis-server mysql-server mysql-client python-dev python libmysqlclient-dev python-pip git libffi-dev  supervisor libssl-dev libtiff5-dev libjpeg8-dev zlib1g-dev libfreetype6-dev liblcms2-dev libwebp-dev -y
 
 
 cd /usr/share/nginx
@@ -53,24 +53,44 @@ fi
 .env/bin/python manage.py db upgrade
 .env/bin/python manage.py init_data
 
-#configure UWSGI
-mkdir -p /etc/uwsgi
-mkdir -p /etc/uwsgi/vassals
-rm -rf /etc/uwsgi/vassals/uwsgi.ini
-ln -s /usr/share/nginx/poppet/scripts/uwsgi.ini /etc/uwsgi/vassals/uwsgi.ini
+if [  -d /etc/init/uwsgi.conf ]; then
+    echo '-------------Upstart UWSGI config found, removing and upgrading to supervisor--------------'
+    if service --status-all | grep -Fq 'uwsgi'; then
+      sudo service uwsgi stop
+    fi
+    rm -rf /etc/init/uwsgi.conf
+
+fi
+
+if [ ! -d /etc/uwsgi ]; then
+    echo '-------------------------Creating /etc/uwsgi-----------------------------------------'
+    mkdir /etc/uwsgi
+
+fi
+if [ ! -d /etc/uwsgi/vassals ]; then
+    echo '-------------------------Creating /etc/uwsgi/vassals-----------------------------------------'        
+    mkdir /etc/uwsgi/vassals
+    ln -s /usr/share/nginx/poppet/scripts/uwsgi.ini /etc/uwsgi/vassals/uwsgi.ini
+
+fi
+
+if [ ! -f /etc/uwsgi/vassals/uwsgi.ini ]; then
+    echo '-------------------------Configuring /etc/uwsgi/vassals/uwsgi.ini-----------------------------'
+    ln -s /usr/share/nginx/poppet/scripts/uwsgi.ini /etc/uwsgi/vassals/uwsgi.ini
+
+fi
+if [ ! -f /etc/nginx/sites-enabled/wifiapp ]; then
+    echo '-------------------------Configuring nGINX-----------------------------------------'
+    rm -rf /etc/nginx/sites-enabled/default
+    cp /usr/share/nginx/poppet/scripts/wifiapp.conf /etc/nginx/sites-available/wifiapp.conf
+    ln -s /etc/nginx/sites-available/wifiapp.conf /etc/nginx/sites-enabled/wifiapp
+fi
+
+if [ ! -f /etc/supervisor/conf.d/unifispot_celery.conf ]; then
+    echo '-------------------------Configuring celeryd-----------------------------------------'
+    ln -s /usr/share/nginx/poppet/scripts/supervisord.conf  /etc/supervisor/conf.d/unifispot_celery.conf
+fi
+
 chown -R www-data:www-data /usr/share/nginx/poppet
-cp uwsgi.conf /etc/init/
-
-#configure Nginx
-rm -rf /etc/nginx/sites-enabled/default
-cp scripts/wifiapp.conf /etc/nginx/sites-available/wifiapp.conf
-ln -s /etc/nginx/sites-available/wifiapp.conf /etc/nginx/sites-enabled/wifiapp
-
-service uwsgi restart
 service nginx restart
-
-#configure celery
-apt-get install supervisor
-rm -f  /etc/supervisor/conf.d/poppet_celery.conf
-ln -s /usr/share/nginx/poppet/scripts/supervisord.conf  /etc/supervisor/conf.d/unifispot_celery.conf
 service supervisor restart
